+++ CHANGES.md
@@ -10,6 +10,7 @@
 - NEW: no more dependence with FSL
 - NEW: no more dependence with c3d
 - NEW: **sct_label_utils**: new functionality for creating labels based on vertebral labeling
+- NEW: added "-qc" flag to some functions to output png images for quality control.
 - BUG: **install_patch**: now possible to install as non-admin (issues #380, #434)
 - OPT: **sct_straighten_spinalcord**: improved accuracy (issues #371, #425, #452, #472)